OSIS-428 move enum offer_enrollment_state.py attribution/models/enums/offer_enrollment_state.py -> base/models/enums/offer_enrollment_state.py
